+++ src/code/util/reexports.rs
@@ -7,7 +7,7 @@
 
 use crate::reexport;
 
-pub use devela_base::{items, sf};
+pub use devela_base::{CONST, is, items, sf};
 
 /// <span class="stab portability" title="re-exported from the `devela_macros`
 /// crate">`devela_macros`</span>